Catch: Put Implementation Separate
Put main functions that pull most implementations
into separate .cpp file.

Can also be used as preparation for PCHs.
