+++ test/AuxiliaryTest.cpp
@@ -1,5 +1,3 @@
-#define CATCH_CONFIG_MAIN
-
 // expose private and protected members for invasive testing
 #if openPMD_USE_INVASIVE_TESTS
 #   define OPENPMD_private public
@@ -21,6 +19,7 @@
 #include <fstream>
 #include <string>
 #include <vector>
+#include <iostream>
 
 using namespace openPMD;
