+++ test/CatchMain.cpp
@@ -1,0 +1,4 @@
+#define CATCH_CONFIG_MAIN
+#include <catch2/catch.hpp>
+
+
